Problem: Autodiscovered devices are assigned unexpected inames.
Solution: Type/subtype, which is part of the iname, is read into unnamed std::string object and stored as a char pointer.
This string object is freed after the execution leaves its scope. When we try to use the pointer, memory it points to already contains different data.
To fix the problem, we replace char pointer with named std::string on the uppermost level, which isn't freed until the end of the function.

Signed-off-by: Jana Rapava <janarapava@eaton.com>
(cherry picked from commit 4a42778a57ae8553b3f8fb0cc36ab58a4b5b3eee)